refactor: 구분자 설정 관련 Enum 디렉토리 변경
