+++ src/main/java/calculator/config/CustomDelimiterPattern.java
@@ -1,4 +1,4 @@
-package calculator.model.delimiter;
+package calculator.config;
 
 public enum CustomDelimiterPattern {
 	START("//"),
