+++ src/main/java/calculator/model/delimiter/Delimiters.java
@@ -6,6 +6,7 @@ import java.util.HashSet;
 import java.util.Set;
 import java.util.regex.Pattern;
 
+import calculator.config.DefaultDelimiter;
 import calculator.validation.InputValidator;
 
 public class Delimiters {
